feat: Add class VectorAttributeSchema; Make TigerGraphConnectionConfig support more authentication methods
