+++ tigergraphx/config/graph_db/schema/data_type.py
@@ -1,0 +1,15 @@
+from enum import Enum
+
+
+class DataType(Enum):
+    """
+    Enumeration of supported data types.
+    """
+
+    INT = "INT"
+    UINT = "UINT"
+    FLOAT = "FLOAT"
+    DOUBLE = "DOUBLE"
+    BOOL = "BOOL"
+    STRING = "STRING"
+    DATETIME = "DATETIME"
